Fix log-to-slack script
The issue was that the bash functions only read one line,
and that the pipe was not connected between stdin of the
bash functions and stdin of slacktee, so if we read multiple
lines they would be passed line-by-line to slacktee.

Just solves all this by removing the bash functions,
as they are really not needed.